+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.10.0 (Unreleased)
 
 ENHANCEMENTS:
+
 * Refactoring vpc logic with api3.0
 
 FEATURES:
@@ -11,6 +12,7 @@ FEATURES:
 * **Update Resource**: `tencentcloud_vpc` , add more configurable items.
 * **Update Resource**: `tencentcloud_subnet` , add more configurable items.
 * **Update Resource**: `tencentcloud_route_table` ,add more configurable items.
+* **Update Resource**: `resource/tencentcloud_instance`: add optional argument `tags`.
 
 
 ## 1.9.1 (June 24, 2019)